+++ frontend/src/app/page.tsx
@@ -1,8 +1,42 @@
 import { ProposalTable } from "../components/proposalTable";
 import ScrapeButton from "../components/scrapeButton";
 import { useSession } from "next-auth/react";
+import { IProject } from "@/database/projectSchema";
+// import { GetServerSideProps } from "next";
 
-export default function Home() {
+async function getData(): Promise<IProject[]> {
+  // Fetch data from your API here.
+  try {
+    const response = await fetch("http://localhost:3000/api/projects");
+    if (!response.ok) {
+      throw new Error("Failed to fetch projects");
+    }
+    const data = await response.json();
+    const reformattedProjects = data.map((project: any) => ({
+      countyFileNumber: project.county_file_number,
+      hearingDate: project.hearing_date,
+      reviewStatus: project.review_status ?? "Unreviewed",
+      location: project.location,
+      apn: project.apn,
+      dateAccepted: project.date_accepted,
+      requestingParty: project.requesting_party ?? "N/A",
+      schNumber: project.sch_number ?? "N/A",
+      title: project.title ?? "N/A",
+      publicHearingAgenda: project.public_hearing_agenda_link,
+      schLink: project.sch_page_link ?? "N/A",
+      additionalNotes: project.additonal_notes ?? "N/A",
+    }));
+    return reformattedProjects;
+  } catch (error) {
+    console.error("Error fetching projects:", error);
+    return [];
+  }
+}
+
+export default async function Home() {
+  const proposals = await getData();
+  const numProjects = proposals.length;
+
   return (
     <div className="flex flex-col w-ful gap-5">
       <div className="flex flex-col items-center justify-center bg-primary-foreground py-3 px-40 rounded-lg gap-3">
@@ -19,7 +53,7 @@ export default function Home() {
         <div className="flex justify-center">
           <ScrapeButton />
         </div>
-        <ProposalTable />
+        <ProposalTable data={proposals} numProjects={numProjects} />
       </div>
     </div>
   );
